+++ packages/next/src/client/components/router-reducer/reducers/server-action-reducer.ts
@@ -360,12 +360,17 @@ export function serverActionReducer(
       }
 
       if (redirectLocation && redirectHref) {
-        // Because the RedirectBoundary will trigger a navigation, we need to seed the prefetch cache
-        // with the FlightData that we got from the server action for the target page, so that it's
-        // available when the page is navigated to and doesn't need to be re-fetched.
-        // We only do this if the server action didn't revalidate any data, as in that case the
-        // client cache will be cleared and the data will be re-fetched anyway.
-        if (!actionRevalidated) {
+        if (!process.env.__NEXT_CLIENT_SEGMENT_CACHE && !actionRevalidated) {
+          // Because the RedirectBoundary will trigger a navigation, we need to seed the prefetch cache
+          // with the FlightData that we got from the server action for the target page, so that it's
+          // available when the page is navigated to and doesn't need to be re-fetched.
+          // We only do this if the server action didn't revalidate any data, as in that case the
+          // client cache will be cleared and the data will be re-fetched anyway.
+          // NOTE: We don't do this in the Segment Cache implementation.
+          // Dynamic data should never be placed into the cache, unless it's
+          // "converted" to static data using <Link prefetch={true}>. What we
+          // do instead is re-prefetch links and forms whenever the cache is
+          // invalidated.
           createSeededPrefetchCacheEntry({
             url: redirectLocation,
             data: {
