+++ test/e2e/app-dir/segment-cache/revalidation/components/link-accordion.tsx
@@ -1,6 +1,7 @@
 'use client'
 
 import Link from 'next/link'
+import Form from 'next/form'
 import { useState } from 'react'
 
 export function LinkAccordion({
@@ -27,6 +28,35 @@ export function LinkAccordion({
         </Link>
       ) : (
         <>{children} (link is hidden)</>
+      )}
+    </>
+  )
+}
+
+export function FormAccordion({
+  action,
+  children,
+  prefetch,
+}: {
+  action: string
+  children: React.ReactNode
+  prefetch?: null | false
+}) {
+  const [isVisible, setIsVisible] = useState(false)
+  return (
+    <>
+      <input
+        type="checkbox"
+        checked={isVisible}
+        onChange={() => setIsVisible(!isVisible)}
+        data-form-accordion={action}
+      />
+      {isVisible ? (
+        <Form action={action} prefetch={prefetch}>
+          <button>{children}</button>
+        </Form>
+      ) : (
+        <>{children} (form is hidden)</>
       )}
     </>
   )
